extracted Telnet functions from ComProxy and moved them to a new class TelnetProxy  ComProxy_ut is working: ToDo:  fix make TelnetProxy_ut  fix make x86_client  fix make android_client
